Добавил операторы <, >, = для PrimitiveTest Убрал Generate() из конструкторов Если не была вызвана функция Generate(), то нельзя вызывать функцию Print() Добавил многопоточность

+++ Course_Work/Utils.h
@@ -85,12 +85,13 @@ class RNG
 	static std::function<int()> rng_function_;
 public:
 	RNG();
-	static int Rand();
+	static long long Rand();
 	static void RandomFunction(std::function<int()> _rng_function);
 	static void RandomSeed(unsigned int _seed);
 };
-inline int Rand()
+inline long long Rand()
 {
 	return RNG::Rand();
 }
+
 #endif
